refactor(admin-ui):use-version-info from component to hook
